Energy_Management/requirement.txt
requests
pymysql
python-dotenv
opencv-python
opencv-python-headless

